address comments. rename background_tasks.rs to watermark_task.rs. Consolidate watermark structs to that file. Minimize lock holds by dropping asap
