+++ crates/sui-graphql-rpc/src/server/mod.rs
@@ -3,6 +3,6 @@
 
 pub mod graphiql_server;
 
-pub(crate) mod background_tasks;
 pub mod builder;
 pub mod version;
+pub(crate) mod watermark_task;
